+++ ldaptor/test/test_encoder.py
@@ -6,7 +6,7 @@ from twisted.trial import unittest
 
 import six
 
-from ldaptor.encoder import to_bytes
+from ldaptor.encoder import to_bytes, WireStrAlias
 
 
 class WireableObject(object):
@@ -27,3 +27,10 @@ class EncoderTests(unittest.TestCase):
     def test_bytes_object(self):
         obj = b'bytes'
         self.assertEqual(to_bytes(obj), b'bytes')
+
+
+class WireStrAliasTests(unittest.TestCase):
+
+    def test_not_implemented(self):
+        obj = WireStrAlias()
+        self.assertRaises(NotImplementedError, obj.toWire)